pyproject.toml
[project]
name = "kind-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "pyyaml>=6.0.2",
]
